backend/requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.0.5
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.2.0
idna==3.4
mysqlclient==2.2.0
oauthlib==3.2.2
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.2
tzdata==2023.3
urllib3==1.25.1
openai==0.28.1
